Fix 7 lower-confidence bugs/gaps flagged during the AO3 audit
- migrateMovedEpubState() now retries the cache-dir rename once (the
  main transient cause -- SD contention right after the epub file's
  own rename just succeeded) and, if it still fails, bails before
  touching bookmarks/clippings/recents. Ao3ArchiveUtils::archiveFic/
  restoreFic roll back the epub file move on that failure instead of
  logging it as non-fatal and leaving progress/the AO3 sidecar
  orphaned under a path nothing points at anymore.
- BookActionActivity's Mark-for-Later action checked addBook()'s
  return value nowhere, so hitting the 10-entry cap silently did
  nothing with no feedback; now shows a toast and only flags the
  change when it actually took effect.
- ViewEntry::addedSequence truncated CompactIndexRecord's uint32_t
  counter to uint16_t, which could misorder "Date Added" after 65536
  lifetime index writes on one device. Purely in-RAM (rebuilt from
  the on-disk record every load), so widening it is a free fix.
- Ao3Librarian::writeIndexRecord()'s second header read didn't check
  return values, unlike every sibling read in this file; a transient
  short read could drive the update/append loop off a garbage
  recordCount.
- Ao3WipsStore is deliberately unbounded (every author-incomplete
  fic, not a curated recent-N list), so capping it would silently
  drop real data. Added a heap-floor check on new-entry growth
  instead, matching this codebase's established guard pattern.
- FileBrowserActivity's directory-menu switch had no BookInfo case
  and no default, unlike the file-menu switch 200 lines below it --
  unreachable today, but lost the compiler safety net its sibling has.
- Ao3LibraryActivity::buildAllowedHashes() had its recursion-depth
  guard commented out entirely. Restoring the original guard as-is
  would have broken real matches (deeper folders' epubs are supposed
  to count), so this adds a far looser absolute cap instead, purely
  as a safety net against pathological nesting rather than a
  functional depth limit.

Found via the same 4-branch parallel-agent audit as the previous
commit; these were reported as lower-confidence/unverified rather
than confirmed, and fixed now on request.

+++ src/Ao3ViewEntry.h
@@ -22,15 +22,18 @@ inline uint32_t fnv1a(const char* str) {
  * @brief In-RAM sort/filter key struct — one per live book, loaded sequentially
  *        from ao3_library_index.bin at library startup.
  *
- * 42 bytes packed (pragma pack 1).
- * 42 × 1000 books = 42 KB peak RAM.
+ * 44 bytes packed (pragma pack 1).
+ * 44 × 1000 books = 44 KB peak RAM.
  */
 #pragma pack(push, 1)
 struct ViewEntry {
   uint64_t cacheHash;      // same as CompactIndexRecord.cacheHash
   uint32_t wordCount;      // word count sort
   uint32_t seriesHash;     // fnv1a(seriesName), 0 if no series
-  uint16_t addedSequence;  // date-added sort (monotonic, higher = newer), max 65535
+  uint32_t addedSequence;  // date-added sort (monotonic, higher = newer); matches
+                           // CompactIndexRecord's own width so it can't wrap
+                           // around on a device with a long archive/restore/
+                           // reindex history
   uint16_t seriesPart;     // position within series, 0 if not in a series
   char title[12];          // first 11 chars of title, null-terminated (alphabetic sort)
   char authorKey[8];       // first 7 chars of author lowercased (author sort)
@@ -56,7 +59,7 @@ inline ViewEntry buildViewEntry(const CompactIndexRecord& rec) {
   }
 
   v.wordCount = rec.wordCount;
-  v.addedSequence = static_cast<uint16_t>(rec.addedSequence);
+  v.addedSequence = rec.addedSequence;
   v.seriesHash = fnv1a(rec.seriesName);
   v.seriesPart = rec.seriesPart;
   v.cacheHash = rec.cacheHash;
